Changes to sending mmanager - thread / Changes to executes on psycopg2 / install script

+++ src/database_ops/card_ops.py
@@ -1,10 +1,5 @@
-import psycopg2
-import time
-import datetime
-from flashcard import Card
-from database_ops.db_utils import treat_str_SQL, create_card_with_row
+from database_ops.db_utils import create_card_with_row
 import database_ops.archive_ops
-
 
 
 class CardOps():
@@ -16,7 +11,7 @@ class CardOps():
 		self.archive_ops = database_ops.archive_ops.ArchiveOps(conn,cursor, debug_mode)
 
 	def get_highest_card_id(self, user_id):
-		self.cursor.execute("SELECT highest_card_id FROM users WHERE id={}".format(user_id))
+		self.cursor.execute("SELECT highest_card_id FROM users WHERE id=%s", (user_id,))
 		row = self.cursor.fetchall()
 		if len(row) == 0:
 			return "User doesn't exist"
@@ -34,17 +29,17 @@ class CardOps():
 		topic = card.get_topic()
 
 		#Update user's highest_card_id
-		self.cursor.execute("SELECT highest_card_id FROM users WHERE id={};".format(user_id))
+		self.cursor.execute("SELECT highest_card_id FROM users WHERE id=%s;", (user_id,))
 		highest_card_id = self.cursor.fetchall()
 		highest_card_id = highest_card_id[0][0]
 		
 		if highest_card_id < card_id:
-			self.cursor.execute("UPDATE users SET highest_card_id={} WHERE id={}".format(card_id, user_id))
+			self.cursor.execute("UPDATE users SET highest_card_id=%s WHERE id=%s", (card_id, user_id))
 
 
-		self.cursor.execute("INSERT INTO cards VALUES ({}, {}, '{}', '{}', '{}', {}, '{}', {}, {}, {}, '{}')"
-			.format(user_id, word_id, treat_str_SQL(language), treat_str_SQL(topic), treat_str_SQL(foreign_word),
-					card_id, treat_str_SQL(content_type),
+		self.cursor.execute("INSERT INTO cards VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)",
+			 (user_id, word_id, language, topic, foreign_word,
+					card_id, content_type,
 					card.attempts, card.ef, card.interval,
 				str(card.next_date.year) + '-' + str(card.next_date.month) + '-' + str(card.next_date.day)))
 
@@ -53,15 +48,15 @@ class CardOps():
 		counter = 0
 		for path in card.archives:
 			counter += 1
-			self.cursor.execute("INSERT INTO archives VALUES ({}, {}, {}, '{}', '{}')"
-									.format(user_id, card_id, counter, treat_str_SQL(content_type), treat_str_SQL(path)))
+			self.cursor.execute("INSERT INTO archives VALUES (%s, %s, %s, %s, %s)", 
+									(user_id, card_id, counter, content_type, path))
 
 		self.conn.commit()
 
 
 	def get_card(self, user_id, user_card_id):
-		self.cursor.execute("SELECT * FROM cards WHERE user_id={} AND user_card_id={}"
-				.format(user_id, user_card_id))
+		self.cursor.execute("SELECT * FROM cards WHERE user_id=%s AND user_card_id=%s", 
+							(user_id, user_card_id))
 		row = self.cursor.fetchall()
 
 		if len(row) == 0:
@@ -72,15 +67,15 @@ class CardOps():
 		row = row[0]
 		card = create_card_with_row(row)
 
-		self.cursor.execute("SELECT content_path FROM archives WHERE user_id={} AND user_card_id={};".format(user_id, user_card_id))
+		self.cursor.execute("SELECT content_path FROM archives WHERE user_id=%s AND user_card_id=%s;", (user_id, user_card_id))
 		rows = self.cursor.fetchall()
 		for row in rows:
 			card.add_archive(row[0])
 		return card
 
 	def get_cards_on_topic(self, user_id, language, topic, get_default):
-		self.cursor.execute("SELECT * FROM cards WHERE user_id={} AND language='{}' AND topic='{}'"
-				.format(user_id, treat_str_SQL(language), treat_str_SQL(topic)))
+		self.cursor.execute("SELECT * FROM cards WHERE user_id=%s AND language=%s AND topic=%s", 
+				(user_id, language, topic))
 		rows = self.cursor.fetchall()
 
 		if len(rows) == 0:
@@ -94,7 +89,7 @@ class CardOps():
 			if card.get_type() == 'default' and  get_default == False:
 					continue
 			user_card_id = card.get_card_id()
-			self.cursor.execute("SELECT content_path FROM archives WHERE user_id={} AND user_card_id={};".format(user_id, user_card_id))
+			self.cursor.execute("SELECT content_path FROM archives WHERE user_id=%s AND user_card_id=%s;", (user_id, user_card_id))
 			archives = self.cursor.fetchall()
 			for archive in archives:
 				card.add_archive(archive[0])
@@ -103,8 +98,7 @@ class CardOps():
 		return cards
 
 	def get_all_cards(self, user_id):
-		self.cursor.execute("SELECT * FROM cards WHERE user_id={}"
-				.format(user_id))
+		self.cursor.execute("SELECT * FROM cards WHERE user_id=%s", (user_id, ))
 		rows = self.cursor.fetchall()
 
 		if len(rows) == 0:
@@ -115,7 +109,7 @@ class CardOps():
 		for row in rows:
 			card = create_card_with_row(row)
 			user_card_id = card.get_card_id()
-			self.cursor.execute("SELECT content_path FROM archives WHERE user_id={} AND user_card_id={};".format(user_id, user_card_id))
+			self.cursor.execute("SELECT content_path FROM archives WHERE user_id=%s AND user_card_id=%s;", (user_id, user_card_id))
 			archives = self.cursor.fetchall()
 			for archive in archives:
 				card.add_archive(archive[0])
@@ -125,22 +119,21 @@ class CardOps():
 	
 
 	def erase_card(self, user_id, user_card_id):
-		self.cursor.execute("SELECT * FROM cards WHERE user_id={} AND user_card_id={}"
-						.format(user_id, user_card_id))
+		self.cursor.execute("SELECT * FROM cards WHERE user_id=%s AND user_card_id=%s",
+						(user_id, user_card_id))
 		rows = self.cursor.fetchall()
 		if len(rows) == 0:
 			print("Error in erase_card, dbapi")
-			print("Card {}, {} doesn't exist".format(user_id, user_card_id))
-			return "Card {}, {} doesn't exist".format(user_id, user_card_id)
+			print("Card %s, %s doesn't exist".format(user_id, user_card_id))
+			return "Card %s, %s doesn't exist".format(user_id, user_card_id)
 
 		# erasing archives of the card
-		self.cursor.execute("SELECT counter FROM archives WHERE user_id={} AND user_card_id={};".format(user_id, user_card_id))
+		self.cursor.execute("SELECT counter FROM archives WHERE user_id=%s AND user_card_id=%s;", (user_id, user_card_id))
 		rows = self.cursor.fetchall()
 		for row in rows:
 			self.archive_ops.erase_archive(user_id,user_card_id,row[0])
 
-		self.cursor.execute("DELETE FROM cards WHERE user_id={} AND user_card_id={}"
-						.format(user_id, user_card_id))
+		self.cursor.execute("DELETE FROM cards WHERE user_id=%s AND user_card_id=%s", (user_id, user_card_id))
 		self.conn.commit()
 		return "Card successfuly removed"
 
@@ -152,13 +145,14 @@ class CardOps():
 		word: A Word instance.
 		"""
 		if self.check_card_existence(card.get_user(), card):
-			self.cursor.execute("UPDATE cards SET attempts={}, easiness_factor={}, interval={}, next_date='{}' WHERE user_id={} AND user_card_id={};"
-			.format(card.attempts, card.ef, card.interval, card.next_date.strftime('%Y-%m-%d'), card.get_user(), card.card_id))
+			self.cursor.execute("UPDATE cards SET attempts=%s, easiness_factor=%s, interval=%s, next_date=%s WHERE user_id=%s AND user_card_id=%s;",
+								(card.attempts, card.ef, card.interval, card.next_date.strftime('%Y-%m-%d'), card.get_user(), card.card_id))
 			self.conn.commit()
 
 
 	def check_card_existence(self, user_id, card):
-		self.cursor.execute("SELECT user_card_id FROM cards WHERE user_id={} AND user_card_id={};".format(user_id, card.get_card_id()))
+		self.cursor.execute("SELECT user_card_id FROM cards WHERE user_id=%s AND user_card_id=%s;", 
+																		(user_id, card.get_card_id()))
 		card = self.cursor.fetchall()
 		if len(card) == 0:
 			return False
